committing before something gets effed up, even though textures are kinda wonky now

+++ src/shaders/block.frag
@@ -1,7 +1,7 @@
 #version 450
 
 layout(location=0) in vec2 v_tex_coords;
-layout(location=1) flat in uint v_layer;
+layout(location=1) flat in float v_layer;
 
 layout(location=0) out vec4 f_color;
